updated context.md; cleaned the code
