+++ src/@types/leaflet-draw/index.d.ts
@@ -1,6 +1,8 @@
 // Type definitions for leaflet-draw 0.4
 // Project: https://github.com/Leaflet/Leaflet.draw
-// Definitions by: Matt Guest <https://github.com/matt-guest>, Ryan Blace <https://github.com/reblace>
+// Definitions by: Matt Guest <https://github.com/matt-guest>
+//                 Ryan Blace <https://github.com/reblace>
+//                 Yun Shi <https://github.com/YunS-Stacy>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 // TypeScript Version: 2.3
 
@@ -251,15 +253,52 @@ declare module 'leaflet' {
 			const EDITSTOP: string;
 			const DELETESTART: string;
 			const DELETESTOP: string;
-			
-		}
-		// edited
-		class Polygon {
-			constructor(map: L.Map, options?: L.DrawOptions.PolygonOptions)
 		}
 
-		class CircleMarker {
-			constructor(map: L.Map, options?: L.DrawOptions.CircleOptions)
+		class Feature extends Handler { }
+
+		class SimpleShape extends Feature { }
+
+		class Marker extends Feature {
+			constructor(
+				map: Map,
+				options?: DrawOptions.MarkerOptions
+			)
+		}
+
+		class CircleMarker extends Feature {
+			constructor(
+				map: Map,
+				options?: DrawOptions.MarkerOptions
+			)
+		}
+
+		class Circle extends Feature {
+			constructor(
+				map: Map,
+				options?: DrawOptions.CircleOptions
+			)
+		}
+
+		class Polyline extends Feature {
+			constructor(
+				map: Map,
+				options?: DrawOptions.PolylineOptions
+			)
+		}
+
+		class Rectangle extends Feature {
+			constructor(
+				map: Map,
+				options?: DrawOptions.RectangleOptions
+			)
+		}
+
+		class Polygon extends Feature {
+			constructor(
+				map: Map,
+				options?: DrawOptions.PolygonOptions
+			)
 		}
 	}
