+++ src/app/app.component.ts
@@ -9,7 +9,4 @@ import { Component } from '@angular/core';
 `
 })
 
-export class AppComponent {
-
-  constructor() { }
-}
+export class AppComponent { }
