+++ src/app/pages/map-view/sidebar-section/layer-card/param-item/param-item.component.html
@@ -1,31 +1,29 @@
-<!-- <ng-container *ngIf="type==='number'"> -->
-<label class="checkbox-label" [class.-inactive]="check.checked" [class.-hide]="type!=='number'">
+<label class="checkbox-label" [class.-inactive]="check.checked" [class.-hide]="type !== 'number'">
         <input #check class="checkbox" type="checkbox" (click)="neglectParam(desc.value, num.value)">                       
         <span class="icon"></span>
       </label>
-<select #desc class="desc" [class.-hide]="type!=='number'" (change)="updateValue(desc.value, num.value)">
+<select #desc class="desc" [class.-hide]="type !== 'number'" (change)="updateValue(desc.value, num.value)">
         <option value="1" [selected]="value>0">Near</option>
         <option value="-1" [selected]="value<0">Far</option>
       </select>
-<!-- </ng-container> -->
 
 <span class="name">{{param | rejoin:"_" | titlecase}}</span>
-<select #num [class.-hide]="type!=='number'" class="num" (change)="updateValue(desc.value, num.value)">
-  <option *ngFor="let val of genRange(range)" [value]="val" [selected]="val===this.abs(this.value)">{{val}}</option>
+<select #num [class.-hide]="type !== 'number'" class="num" (change)="updateValue(desc.value, num.value)">
+  <option *ngFor="let val of genRange(range)" [value]="val" [selected]="val === this.abs(this.value)">{{val}}</option>
 </select>
 
-<ng-container [class.-hide]="type!=='string'">
-    <select *ngIf="value=='tin' || value=='idw'" #string class="string" (change)="updateValue(string.value)">
-      <option *ngFor="let val of ['tin','idw']" [value]="val" [selected]="val===value">{{val | uppercase}}</option>
+<ng-container [class.-hide]="type !== 'string'">
+  <select *ngIf="value === 'tin' || value === 'idw'" #string class="string" (change)="updateValue(string.value)">
+      <option *ngFor="let val of ['tin','idw']" [value]="val" [selected]="val === value">{{val | uppercase}}</option>
   </select>
 
 
-    <select *ngIf="value==='hillshade' || value==='png'" #string class="string" (change)="updateValue(string.value)">
-      <option *ngFor="let val of ['hillshade','png']" [value]="val" [selected]="val===value">{{val==='hillshade'?'Hillshade':'Color Ramp'}}</option>
+  <select *ngIf="value === 'hillshade' || value === 'png'" #string class="string" (change)="updateValue(string.value)">
+      <option *ngFor="let val of ['hillshade', 'png']" [value]="val" [selected]="val === value">{{val === 'hillshade' ? 'Hillshade' : 'Color Ramp'}}</option>
   </select>
 
 
-    <select *ngIf="value==='mar10' || value==='jul10'" #string class="string" (change)="updateValue(string.value)">
-      <option *ngFor="let val of ['mar10','jul10']" [value]="val" [selected]="val===value">{{val==='mar10'?'Snow On':'Snow Off'}}</option>
+  <select *ngIf="value === 'mar10' || value === 'jul10'" #string class="string" (change)="updateValue(string.value)">
+      <option *ngFor="let val of ['mar10', 'jul10']" [value]="val" [selected]="val === value">{{val === 'mar10' ? 'Snow On' : 'Snow Off'}}</option>
   </select>
 </ng-container>
