+++ src/app/pages/map-view/sidebar-section/sidebar-header/sidebar-header.component.ts
@@ -1,19 +1,19 @@
 import { Component, EventEmitter, HostBinding, Input, Output, OnInit } from '@angular/core';
 import { Location } from '@angular/common';
 @Component({
-    selector: 'gd-sidebar-header',
-    templateUrl: './sidebar-header.component.html'
+  selector: 'gd-sidebar-header',
+  templateUrl: './sidebar-header.component.html'
 })
 export class SidebarHeaderComponent {
-    @HostBinding('class.-collapsed') @Input() isCollapsed = false;
-    @Output() isCollapsedChange = new EventEmitter<boolean>();
-    @Input() title: string;
+  @HostBinding('class.-collapsed') @Input() isCollapsed = false;
+  @Output() isCollapsedChange = new EventEmitter<boolean>();
+  @Input() title: string;
 
-    goBack() {
-        this.location.back();
-    }
+  goBack() {
+    this.location.back();
+  }
 
-    constructor(
-        private location: Location
-    ) { }
+  constructor(
+    private location: Location
+  ) { }
 }
