+++ src/app/shared/directives/action-btn.directive.ts
@@ -1,4 +1,4 @@
-import { Directive, Input, Output, EventEmitter, HostBinding, HostListener } from '@angular/core';
+import { Directive, Input, Output, EventEmitter, HostListener } from '@angular/core';
 
 @Directive({
   selector: '[gdActionBtn]'
@@ -6,12 +6,9 @@ import { Directive, Input, Output, EventEmitter, HostBinding, HostListener } fro
 export class ActionBtnDirective {
   @Input() action: string;
   @Input() expanded: string;
-
   @Output() expandedChange = new EventEmitter<string>();
   @HostListener('click') onClick() {
     this.expanded = this.expanded === this.action ? undefined : this.action;
     this.expandedChange.emit(this.expanded);
   }
-  constructor() { }
-
 }
