+++ src/app/shared/pipes/rejoin.pipe.ts
